feat: shared MaterialBadge component, sticky table headers
- Extract MaterialBadge into shared stat-display component
- Use colored material badges on workshops page, workshop detail, and material cards
- Remove duplicate badge code from crafting page
- Add sticky table headers with scrollable container
