+++ src/pages/crafting/crafting-page.tsx
@@ -33,7 +33,7 @@ import {
   SelectTrigger,
   SelectValue,
 } from "@/components/ui/select"
-import { StatDisplay } from "@/components/stat-display"
+import { MaterialBadge, StatDisplay } from "@/components/stat-display"
 import {
   gameApi,
   type CraftingRecipe,
@@ -1074,16 +1074,6 @@ function ReverseTable({
   )
 }
 
-const MAT_BADGE_COLORS: Record<string, string> = {
-  Wood: "bg-amber-900/60 text-amber-200 border-amber-700/50",
-  Leather: "bg-amber-700/60 text-amber-100 border-amber-600/50",
-  Bronze: "bg-orange-600/60 text-orange-100 border-orange-500/50",
-  Iron: "bg-slate-500/60 text-slate-100 border-slate-400/50",
-  Hagane: "bg-blue-600/60 text-blue-100 border-blue-500/50",
-  Silver: "bg-gray-300/70 text-gray-900 border-gray-400/50",
-  Damascus: "bg-purple-600/60 text-purple-100 border-purple-500/50",
-}
-
 function SlotCell({
   name,
   type,
@@ -1131,17 +1121,4 @@ function SortIcon({ sorted }: { sorted: false | "asc" | "desc" }) {
   if (sorted === "asc") return <ArrowUp className="size-3.5" />
   if (sorted === "desc") return <ArrowDown className="size-3.5" />
   return <ArrowUpDown className="text-muted-foreground/50 size-3.5" />
-}
-
-function MaterialBadge({ mat }: { mat: string }) {
-  return (
-    <span
-      className={cn(
-        "rounded border px-1.5 py-0.5 text-xs font-medium",
-        MAT_BADGE_COLORS[mat] ?? "bg-muted"
-      )}
-    >
-      {mat}
-    </span>
-  )
 }
